pyproject.toml
[tool.poetry]
name = "pixie"
version = "0.0.2"
description = "Used to run pixie scripts."
authors = ["Dan Clayton <dan@azwebmaster.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
Jinja2 = "3.0.3"
PyGithub = "^1.55"
GitPython = "^3.1.24"
"ruamel.yaml" = "^0.17.20"
requests = "^2.27.1"

[tool.poetry.scripts]
pixie = "pixie.cli:cli"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.12b0", allow-prereleases = true}
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v"
